+++ pyarr/radarr_api_v3.py
@@ -21,10 +21,135 @@ class RadarrAPIv3(RequestAPI):
 
     ## Movies
     # TODO: GET Movie
-    # TODO: POST Movie
-    # TODO: PUT Movie
-    # TODO: DELETE Movie
+    def get_movie(self, tmdbid = None):
+        """get_movie returns all movies in collection.
+
+        Args:
+            Optional - tmdbid - TMDb id of the movie to get
+        Returns:
+            json response
+
+        """
+        if tmdbid:
+            path = f"/api/v3/movie/{tmdbid[0]}"
+        else:
+            path = "/api/v3/movie"
+
+        res = self.request_get(path)
+        return res.json()
+
     # TODO: GET Movie Lookup
+    def lookup_movie(self, term):
+        """Searches for movie
+
+        Args:
+            Requried - term (uses tmdb for search results)
+        Returns:
+            json response
+
+        """
+        term = str(term)
+        term = term.replace(" ", "%20")
+        path = f"/api/v3/movie/lookup?term={term}"
+        res = self.request_get(path)
+        return res.json()
+
+    def get_root(self):
+        """Returns the Root Folder"""
+        path = "/api/v3/rootfolder"
+        res = self.request_get(path)
+        return res.json()
+
+    def get_quality_profiles(self):
+        """Query Radarr for quality profiles"""
+        path = "/api/v3/qualityProfile"
+        res = self.request_get(path)
+        return res.json()
+
+    def construct_movie_json(
+        self, dbId, qualityProfileId, rootDir, monitored=True, searchForMovie=True
+    ):
+        """Searches for movie on tmdb and returns Movie json to add
+
+        Args:
+            Required - dbID, <imdb or tmdb id>
+            Required - qualityProfileId (int)
+            Required - rootDir (string)
+            Optional - monitored (boolean)
+            Optional - searchForMovie (boolean)
+
+        Return:
+            JsonArray
+
+        """
+        s_dict = self.lookup_movie(dbId)
+
+        movie_json = {
+            "title": s_dict["title"],
+            "path": rootDir + s_dict["title"],
+            "qualityProfileId": qualityProfileId,
+            "year": s_dict["year"],
+            "tmdbId": s_dict["tmdbId"],
+            "images": s_dict["images"],
+            "titleSlug": s_dict["titleSlug"],
+            "monitored": monitored,
+            "addOptions": {"searchForMovie": searchForMovie},
+        }
+        return movie_json
+
+    # POST Movie
+    def add_movie(
+        self, dbId, qualityProfileId, rootDir, monitored=True, searchForMovie=True
+    ):
+        """addMovie adds a new movie to collection
+
+        Args:
+            Required - dbid <imdb or tmdb id>
+            Required - qualityProfileId (int)
+            Required - rootDir (string)
+        Returns:
+            json response
+
+        """
+        movie_json = self.construct_movie_json(
+            dbId, qualityProfileId, rootDir, monitored, searchForMovie
+        )
+
+        path = "/api/v3/movie"
+        res = self.request_post(path, data=movie_json)
+        return res.json()
+
+    # TODO: PUT Movie
+    def update_movie(self, data):
+        """Update an existing movie.
+
+        Args:
+            data (dictionary containing an object obtained by get_movie())
+        Returns:
+            json response
+        """
+
+        path = "/api/movie"
+        res = self.request_put(path, data)
+        return res.json()
+
+    # DELETE Movie
+    def del_movie(self, movieId, delFiles=False, addExclusion=False):
+        """Delete a single movie by database id
+        Args:
+            Required - movieId (int)
+            Optional - delFiles (bool)
+            Optional - addExclusion (bool)
+        Returns:
+            json response
+
+        """
+        # File deletion does not work
+        data = {"deleteFiles": delFiles, "addExclusion": addExclusion}
+        path = f"/api/v3/movie/{movieId}"
+        res = self.request_del(path, data)
+        return res.json()
+
     # TODO: PUT Movie Editor
     # TODO: DELETE Movie Editor
     # TODO: POST Movie import
@@ -36,7 +161,42 @@ class RadarrAPIv3(RequestAPI):
 
     ## history
     # TODO: GET history
+    def get_history(self, page, **kwargs):
+        """Return a json object list of items in your history
+
+        Args:
+            Required - page (int) - 1-indexed (1 default)
+            Optional - sortKey (string) - movie.title or date
+            Optional - pageSize (int) - Default: 0
+            Optional - sortDir (string) - asc or desc - Default: asc
+        Returns:
+            json response
+        """
+        data = {}
+        data.update({"page": kwargs.get("page", 1)})
+        for key, value in kwargs.items():
+            data.update({key: value})
+        path = "/api/v3/history/"
+        res = self.request_get(path, **data)
+        return res.json()
+
     # TODO: GET History Movies
+    def get_history_movie(self, page, movieId, eventType=None):
+        """Return a json object list of items in your history
+
+        Args:
+            Required - movieId (Database id of movie)
+            Optional - eventType (History event type to retrieve)
+        Returns:
+            json response
+        """
+        data = {"modieId": movieId}
+        if eventType:
+            data.update({"eventType": eventType})
+
+        path = "/api/v3/history/movie"
+        res = self.request_get(path, data)
+        return res.json()
 
     ## blacklist
     # TODO: GET blacklist
@@ -45,7 +205,34 @@ class RadarrAPIv3(RequestAPI):
     # TODO: DELETE Blacklist Bulk
 
     ## queue
-    # TODO: GET Queue
+    # GET
+    def get_queue(self):
+        """Return a json object list of items in the queue"""
+        path = "/api/v3/queue"
+        res = self.request_get(path)
+        return res.json()
+
+    # TODO: DELETE Check still exists
+    def del_queue(self, id, blacklist=None):
+        """Deletes an item from the queue and download client. Optionally blacklist item after deletion.
+
+        Args:
+            Required - id (int)
+            Optional - blacklist (bool)
+        Returns:
+            json response
+        """
+        data = {}
+        data.update({"id": id})
+        if blacklist:
+            data.update(
+                {
+                    "blacklist": blacklist,
+                }
+            )
+        path = "/api/v3/queue/"
+        res = self.request_del(path, data)
+        return res.json()
 
     ## indexer
     # TODO: GET indexer
@@ -62,6 +249,20 @@ class RadarrAPIv3(RequestAPI):
     ## Tag
 
     ## diskspace
+    def get_disk_space(self):
+        """Query Radarr for disk usage information
+
+            Location: System > Status
+
+        Args:
+            None
+        Returns:
+            json response
+
+        """
+        path = "/api/v3/diskspace"
+        res = self.request_get(path)
+        return res.json()
 
     ## Settings
 
@@ -131,6 +332,34 @@ class RadarrAPIv3(RequestAPI):
         return res.json()
 
     ## calendar
+    def get_calendar(self, unmonitored, start_date=None, end_date=None):
+        """Get a list of movies based on calendar parameters.
+        If start and end are not provided, retrieves movies airing today and tomorrow.
+
+         args:
+            Required - unmonitored (bool)
+            Optional -
+                - start_date (datetime ISO 8601):
+                - end_date (datetime ISO 8601):
+
+         Returns:
+             json response
+
+        """
+        path = "/api/v3/calendar"
+        data = {"unmonitored": unmonitored}
+
+        if start_date and end_date:
+            if isinstance(start_date, datetime):
+                startDate = start_date.strftime("%Y-%m-%d")
+                data.update({"start": startDate})
+
+            if isinstance(end_date, datetime):
+                endDate = end_date.strftime("%Y-%m-%d")
+                data.update({"end": endDate})
+
+        res = self.request_get(path, **data)
+        return res.json()
 
     ## custom filters