+++ pyarr/request_api.py
@@ -1,6 +1,5 @@
 import requests
 
-
 class RequestAPI:
     def __init__(
         self,